+++ tests/mocks/tendermint_tendermint_libs_log_DB.go
@@ -5,7 +5,7 @@
 package mocks
 
 import (
-	"reflect"
+	reflect "reflect"
 
 	gomock "github.com/golang/mock/gomock"
 	log "github.com/line/ostracon/libs/log"