+++ x/foundation/keeper/keeper_test.go
@@ -189,9 +189,6 @@ func (s *KeeperTestSuite) SetupTest() {
 	// grant stranger to receive foundation treasury
 	err = s.keeper.Grant(s.ctx, s.stranger, &foundation.ReceiveFromTreasuryAuthorization{})
 	s.Require().NoError(err)
-
-	// set gov-mint left count to 1
-	s.keeper.SetGovMintLeftCount(s.ctx, 1)
 }
 
 func TestKeeperTestSuite(t *testing.T) {
@@ -222,19 +219,20 @@ func TestNewKeeper(t *testing.T) {
 
 	for name, tc := range testCases {
 		tc := tc
+		t.Run(name, func(t *testing.T) {
+			newKeeper := func() keeper.Keeper {
+				app := simapp.Setup(false)
+				return keeper.NewKeeper(app.AppCodec(), sdk.NewKVStoreKey(foundation.StoreKey), app.MsgServiceRouter(), app.AccountKeeper, app.BankKeeper, authtypes.FeeCollectorName, foundation.DefaultConfig(), tc.authority.String())
+			}
 
-		newKeeper := func() keeper.Keeper {
-			app := simapp.Setup(false)
-			return keeper.NewKeeper(app.AppCodec(), sdk.NewKVStoreKey(foundation.StoreKey), app.MsgServiceRouter(), app.AccountKeeper, app.BankKeeper, authtypes.FeeCollectorName, foundation.DefaultConfig(), tc.authority.String())
-		}
+			if tc.panics {
+				require.Panics(t, func() { newKeeper() })
+				return
+			}
+			require.NotPanics(t, func() { newKeeper() })
 
-		if tc.panics {
-			require.Panics(t, func() { newKeeper() }, name)
-			continue
-		}
-		require.NotPanics(t, func() { newKeeper() }, name)
-
-		k := newKeeper()
-		require.Equal(t, authority.String(), k.GetAuthority(), name)
+			k := newKeeper()
+			require.Equal(t, authority.String(), k.GetAuthority())
+		})
 	}
 }