+++ x/foundation/keys.go
@@ -10,6 +10,5 @@ const (
 	// RouterKey defines the module's message routing key
 	RouterKey = ModuleName
 
-	TreasuryName  = "treasury"
-	GovMinterName = "gov_minter"
+	TreasuryName = "treasury"
 )