+++ CHANGELOG.md
@@ -38,15 +38,29 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ## [Unreleased](https://github.com/line/lbm-sdk/compare/v0.46.0...HEAD)
 
 ### Features
+* (baseapp) [\#840](https://github.com/line/lbm-sdk/pull/840) allow querying the state based on `CheckState`.
+* (x/foundation) [\#848](https://github.com/line/lbm-sdk/pull/848) remove `gov mint` for x/foundation proposal
+* (x/wasm) [\#850](https://github.com/line/lbm-sdk/pull/850) remove `x/wasm` module in lbm-sdk
 
 ### Improvements
+* (cosmovisor) [\#792](https://github.com/line/lbm-sdk/pull/792) Use upstream's cosmovisor
+* (server) [\#821](https://github.com/line/lbm-sdk/pull/821) Get validator pubkey considering KMS
 
 ### Bug Fixes
+* (client) [\#817](https://github.com/line/lbm-sdk/pull/817) remove support for composite (BLS) type
+
+* (x/foundation) [#834](https://github.com/line/lbm-sdk/pull/834) Apply foundation audit
+
+### Removed
+* [\#853](https://github.com/line/lbm-sdk/pull/853) remove useless stub BeginBlock, EndBlock methods from modules below
+  * ibc, authz, collection, feegrant, ibc, token, wasm
 
 ### Breaking Changes
+* (rest) [\#807](https://github.com/line/lbm-sdk/pull/807) remove legacy REST API
 
 * (codec) [#833](https://github.com/line/lbm-sdk/pull/833) Fix foundation amino codec
 
 ### Build, CI
+* (ci) [\#829](https://github.com/line/lbm-sdk/pull/829) automate release process
 
 ### Document Updates